Add module for task completion
